% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessBridge.R
\name{access_bridge}
\alias{access_bridge}
\title{Wrapper function to link to data}
\usage{
access_bridge(data_type, conn_string, database, table_name, username, password,
  type = NA)
}
\value{
Returns the linked data (as \code{data.table}), connection string
  (\code{conn_string} for ORE database) and name of database or flat table.
}
\description{
Wrapper function to link to data
}
