% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessBridge.R
\name{check_data_type}
\alias{check_data_type}
\title{Checks whether input data type is supported}
\usage{
check_data_type(data.type)
}
\arguments{
\item{data.type}{\code{flat} for flat tables and \code{sql} fo MySQL database.}
}
\value{
Returns \code{data.type} except for \code{xls} and \code{xlsx}, in
  which case \code{excel} is returned instead because both types of files
  share the same function for importing data.
}
\description{
Checks whether input data type is supported
}
