% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_input}
\alias{check_input}
\title{Check validity of input for the 3 main functions}
\usage{
check_input(...)
}
\description{
Check validity of input for the 3 main functions
}
\details{
This function is needed because default values of input is
  overwritten to \code{""} by Rcmdr.
}
\examples{
{
check_input(NA)
check_input(NULL)
check_input(NULL, NULL)
check_input(NA, NULL)
check_input(NULL, NA)
check_input(NULL, "a")
check_input("a", NULL)
check_input(c("a", NULL, NA, "b", ""))
}
}
