% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_long}
\alias{check_long}
\title{Validate request form in long format}
\usage{
check_long(research_folder, database, request_file, template_file)
}
\description{
Validate request form in long format
}
\details{
\code{request_file} is annotated inclusion list, and
  \code{template_file} is the one generated by \code{genInclusion}. Column
  names and \code{key.var} should be identical in the two files.
}
