% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_key_short}
\alias{compare_key_short}
\title{Write a table to compare key_var and key_desc for short version inclusion}
\usage{
compare_key_short(setting, key)
}
\description{
Write a table to compare key_var and key_desc for short version inclusion
}
