% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_inclu}
\alias{count_inclu}
\title{Count based on inclusion}
\usage{
count_inclu(inclu_list)
}
\arguments{
\item{inclu_list}{A list of inclusion to combine.}
}
\description{
Count based on inclusion
}
