% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract data by users' requirements.}
\usage{
extract_data(wkdir = getwd(), research.folder = NA,
  inclusion.xls.file = NA, variable.xls.file = NA, database = NA,
  dataLogic = NA, select.output = NA, overwrite = TRUE,
  username = NA, password = NA, conn_string = NA)
}
\arguments{
\item{wkdir}{The path to working directory. See \code{\link{initWkdir}} for
details on a working directory.}

\item{research.folder}{The name of research folder. See
\code{\link{initResearchFolder}} for details on a research folder.}

\item{inclusion.xls.file}{Name(s) of request form(s) with inclusion
criterion. Multiple request forms should be specified as a character
vector.}

\item{variable.xls.file}{Name(s) of request form(s) with variable lists.
Multiple request forms should be specified as a character vector. Default
is \code{NA}, where no variable list is specified, and variables in the
inclusion criteria will be extracted instead.}

\item{database}{Name of database. Should be either \code{private} or
\code{public} for flat tables, indicating whether data is stored in
\code{public_data} or \code{research/[research folder]/private_data}, or
the actual name of database if extracting data from a database.}

\item{dataLogic}{Whether to take \code{union} or \code{intersection} of
inclusion criteria, if multiple criteria specified. Default is \code{NA}.}

\item{select.output}{\code{1} to generate lists of identifier variables from
merged inclusion criteria; \code{2} to generate an Excel file with summary
statistics for both inclusion criteria and variable lists; \code{3} to
generate \code{csv} files with data extracted based on each request form;
\code{4} to generate a single \code{csv} file for the final merged data.
Multiple selection should be specified as a vector.}

\item{overwrite}{Whether to overwrite existing request form. Default is
\code{TRUE}.}

\item{username}{User name for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{password}{Password for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{conn_string}{Connection string for accessing ORE server. Default is
\code{NA}.}
}
\value{
Returns a list of identifier variables, path to the Excel file with
  summary statistics, extracted data and merged data, if any of these are
  selected with \code{selected.output}. These are also written as \code{csv}
  files in \code{research/[research folder]/request_output} folder.The
  \code{summary.xls} is returned. The \code{summary.xls} will includes count
  summary sheet and variable summary sheet.
}
\description{
This function extracts data based on the request forms users
  have filled and saved in the \code{request_output} folder of selected
  research folder.
}
\examples{
\dontrun{
extract_data(wkdir = "Working directory", research.folder = "requestnum001",
             inclusion.xls.file = "inclusion.Diagnosis_DIAGNOSIS_CD(DIAGNOSIS_DESC_ICD_VERSION)",
             variable.xls.file = "variable.Patient(PATIENT_NRIC)",
             select.output = c(1, 2, 4))
}
}
\seealso{
\code{\link{genInclusion}}, \code{\link{genVariable}}
}
