% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_key}
\alias{extract_key}
\title{Read in inclusion and variable (wrapper)}
\usage{
extract_key(research_folder, database, file, key_var, key_desc = NA, type)
}
\description{
Read in inclusion and variable (wrapper)
}
