% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_key_long}
\alias{extract_key_long}
\title{Read in inclusion in long format or variable list (in this case key_var is
variable)}
\usage{
extract_key_long(research_folder, database, file, key_var)
}
\description{
Read in inclusion in long format or variable list (in this case key_var is
variable)
}
