% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_key_short}
\alias{extract_key_short}
\title{Read in inclusion in short format}
\usage{
extract_key_short(research_folder, file, key_var, key_desc = NA)
}
\description{
Read in inclusion in short format
}
