% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_selection}
\alias{extract_selection}
\alias{extract_selection_logic}
\title{Extract inclusion criteria based on selection symbols}
\usage{
extract_selection(selection, selection_symbol, key.var)

extract_selection_logic(logic, request, key.var)
}
\arguments{
\item{selection}{The column for selection.}

\item{selection_symbol}{Symbol to indicate selection.}

\item{key.var}{The name of variable based on which inclusion criterion is
specified. Vector with multiple elements is not allowed. To have complex
inclusion criteria based on multiple variables, generate one request form
for each variable.}

\item{logic}{The column for logic.}

\item{request}{Data read from request file.}
}
\description{
Extract inclusion criteria based on selection symbols
}
\section{Functions}{
\itemize{
\item \code{extract_selection_logic}: Extract selection based on logit.
}}

