% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R, R/accessData.R, R/accessDataCol.R
\name{fail_error}
\alias{fail_error}
\alias{fail_error}
\alias{fail_error}
\title{Flag error for failing to find a table}
\usage{
fail_error(table_name, database, is_database = TRUE)

fail_error(table_name, database, is_database = TRUE)

fail_error(table_name, database, is_database = TRUE)
}
\description{
Flag error for failing to find a table

Flag error for failing to find a table

Flag error for failing to find a table
}
