% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_data}
\alias{filter_data}
\title{Filter data based on identifier variable}
\usage{
filter_data(dat, identifier_var, id_merged)
}
\arguments{
\item{dat}{A data.frame.}

\item{identifier_var}{Identifier variable.}

\item{id_merged}{A data.frame of merged identifiers.}
}
\description{
Filter data based on identifier variable
}
