% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_common}
\alias{find_common}
\title{Find common strings in all elements of a list of strings}
\usage{
find_common(s_list)
}
\arguments{
\item{s_list}{A list of strings.}
}
\description{
Find common strings in all elements of a list of strings
}
