% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genInclusion.R
\name{genInclusion}
\alias{genInclusion}
\title{Generate inclusion criterion}
\usage{
genInclusion(
  wkdir = getwd(),
  research.folder = NA,
  table_name = NA,
  data.type = NA,
  database = NA,
  key.var = NA,
  key.desc = NA,
  identifier.var = NA,
  count = NA,
  overwrite = TRUE,
  username = NA,
  password = NA
)
}
\arguments{
\item{wkdir}{The path to working directory. See \code{\link{initWkdir}} for
details on a working directory.}

\item{research.folder}{The name of research folder. See
\code{\link{initResearchFolder}} for details on a research folder.}

\item{table_name}{The name of dataset. File extension must be included if the
data is saved as a flat file instead of in a database, e.g., `data.csv`
should be given instead of simply `data`.}

\item{data.type}{\code{flat} for flat tables and \code{sql} fo MySQL database.}

\item{database}{Name of database. Should be either \code{private} or
\code{public} for flat tables, indicating whether data is stored in
\code{public_data} or \code{research/[research folder]/private_data}, or
the actual name of database if extracting data from a database.}

\item{key.var}{The name of variable based on which inclusion criterion is
specified. Vector with multiple elements is not allowed. To have complex
inclusion criteria based on multiple variables, generate one request form
for each variable.}

\item{key.desc}{The name of variable(s) that describes \code{key.var}, if
any. Multiple description variables should be specified as a character
vector. Default is \code{NA}, where no description variable is selected.}

\item{identifier.var}{The name(s) of identifier variable(s) of the inclusion
criterion. Multiple identifier variables should be specified as a character
vector.}

\item{count}{Whether a count summary should be generated for each of
\code{identifier.var}, as a logical vector with the same length as
\code{identifier.var}. Default value is
\code{rep(TRUE, length(identifier.var))}.}

\item{overwrite}{Whether to overwrite existing request form. Default is
\code{TRUE}.}

\item{username}{User name for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{password}{Password for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}
}
\value{
Returns the name of inclusion criteria generated. See \code{details}.
}
\description{
This function generates a request form containing a list of all
  unique values of a particular key variable in a dataset selected, based on
  which users can specify their inclusion criteria for data extraction.
}
\details{
Writes an \code{.xls} file with two sheets to
  \code{public_data_template} if lining to database, and to either
  \code{public_data_template} or \code{private_data_template} depending on
  \code{database} for flat tables.

The first sheet \code{overall list} contains unique values of key variable
with description (if \code{key.desc} is specfied and exists in the table),
remarks and selection. This sheet will be used to specify inclusion criteria
for data extraction in later steps.

The second sheet \code{setting list} contains detailed information about
the database.
}
\examples{
\dontrun{
# Generate the inclusion criterion from public table `movement_table.txt`
# based on variable `DIAGNOSIS_CD`. `DIAGNOSIS_DESC` and `ICD_VERSION`
# contains detailed description for the key variable.
#
# Unique entries are identified by `PATIENT_NRIC` and `CASE_NO`.
# Unique `DIAGNOSIS_CD` is counted by `PATIENT_NRIC` and `CASE_NO`.
#
genInclusion(wkdir = "Working directory", research.folder = "requestnum001",
             table_name = `movement_table.txt`, data.type = "flat",
             database = "public",
             key.var = "DIAGNOSIS_CD",
             key.desc = c("DIAGNOSIS_DESC", "ICD_VERSION"),
             identifier.var = c("PATIENT_NRIC", "CASE_NO"))
}
}
\seealso{
\code{\link{genVariable}}, \code{\link{extract_data}}
}
