% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genVariable.R
\name{genVariable}
\alias{genVariable}
\title{Generate variable list.}
\usage{
genVariable(
  wkdir = getwd(),
  research.folder = NA,
  table_name = NA,
  data.type = NA,
  database = NA,
  identifier.var = NA,
  omit.var = NA,
  overwrite = TRUE,
  username = NA,
  password = NA
)
}
\arguments{
\item{wkdir}{The path to working directory. See \code{\link{initWkdir}} for
details on a working directory.}

\item{research.folder}{The name of research folder. See
\code{\link{initResearchFolder}} for details on a research folder.}

\item{table_name}{The name of dataset. File extension must be included if the
data is saved as a flat file instead of in a database, e.g. `data.csv`
should be given instead of simply `data`.}

\item{data.type}{\code{flat} for flat tables and \code{sql} fo MySQL database.}

\item{database}{Name of database. Should be either \code{private} or
\code{public} for flat tables, indicating whether data is stored in
\code{public_data} or \code{research/[research folder]/private_data}, or
the actual name of database if extracting data from a database.}

\item{identifier.var}{The name(s) of identifier variable(s) of the variable
list. Multiple identifier variables should be specified as a character
vector.}

\item{omit.var}{The name(s) of variables in the dataset to exclude from the
variable list, if any. \code{identifier.var} are automatically excluded
from the resulting variable list since they will always be in the extracted
data. Multiple omit variables should be specified as a character vector.
Default is \code{NA}.}

\item{overwrite}{Whether to overwrite existing request form. Default is
\code{TRUE}.}

\item{username}{User name for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{password}{Password for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}
}
\value{
Returns the name of variable list generated. See \code{details}.
}
\description{
This function generates a request form containing a list of
  variables available in a particular key variable in a dataset selected,
  based on which users can select variables to extract for data extraction.
}
\details{
Writes an \code{.xls} file with two sheets to
  \code{public_data_template} if lining to database, and to either
  \code{public_data_template} or \code{private_data_template} depending on
  \code{database} for flat tables.

The first sheet \code{overall list} contains unique values of key variable
with description (if \code{key.desc} is specfied and exists in the table),
remarks and selection. This sheet will be used to specify inclusion criteria
for data extraction in later steps.

The second sheet \code{setting list} contains detailed information about
the database.
}
\examples{
\dontrun{
genVariable(wkdir = "Working directory", research.folder = "requestnum001",
            table_name = "table_demographic.txt",
            data.type = "flat", database = "public",
            identifier.var = "PATIENT_NRIC",
            omit.var = c("PATIENT_NRIC", "NATIONALITY", "POSTAL_CODE"))
}
}
\seealso{
\code{\link{genInclusion}}, \code{\link{extract_data}}
}
