% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_directories.R
\name{initResearchFolder}
\alias{initResearchFolder}
\title{Initialise research folder}
\usage{
initResearchFolder(wkdir, research.folder)
}
\arguments{
\item{wkdir}{The working directory of all subfolders, e.g., \code{Working
directory}.}

\item{research.folder}{Name of the research folder.}
}
\description{
Initialise research folder
}
\details{
This function initialises the research folder by automatically
  creating the research folder specified within the \code{research} subfolder
  of the working directory, and then creating the \code{private_data},
  \code{private_data_template}, \code{request_input} and
  \code{request_output} in the research folder. This process can also be done
  manually, as long as the folder structure and folder names are followed
  exactly.

  After the research folder is initialised, user should make sure that all
  private non-database data files are within the \code{private_data}
  subfolder. Relevant inclusion files and variables lists should be saved in
  \code{request_input} subfolder within the research folder for data
  extraction.
}
