% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_directories.R
\name{initWkdir}
\alias{initWkdir}
\title{Initialise working directory}
\usage{
initWkdir(wkdir)
}
\arguments{
\item{wkdir}{The working directory of all subfolders, e.g., \code{Working
directory}.}
}
\description{
Initialise working directory
}
\details{
This function initialises the working directory by automatically
  creating the \code{public_data}, \code{public_template} and \code{research}
  subfolders for users, which can also be manually created by user as long as
  the folder names are followed exactly.

  After the working directory is initialised, user should make sure that all
  public non-database data files are within the \code{public_data} subfolder.
}
