% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R, R/accessData.R, R/accessDataCol.R
\name{list_TN_oracle}
\alias{list_TN_oracle}
\alias{list_TN_mysql}
\alias{list_VN_oracle}
\alias{list_VN_mysql}
\alias{access_oracle}
\alias{access_mysql}
\alias{access_oracle_col}
\alias{access_mysql_col}
\title{Functions to link to Oracle database and extract column name}
\usage{
list_TN_oracle(database, username, password)

list_TN_mysql(database, username, password)

list_VN_oracle(database, table_name, username, password)

list_VN_mysql(database, table_name, username, password)

access_oracle(database, table_name, username, password)

access_mysql(database, table_name, username, password)

access_oracle_col(database, table_name, username, password)

access_mysql_col(database, table_name, username, password)
}
\description{
Functions to link to Oracle database and extract column name

Functions to link to MySQL database and extract column name

Functions to link to Oracle database and extract column name

Functions to link to MySQL database and extract column name

Functions to link to Oracle database

Functions to link to MySQL database

Functions to link to Oracle database and extract column name

Functions to link to MySQL database and extract column name
}
\section{Functions}{
\itemize{
\item \code{list_TN_oracle}: to Oracle database and extract the column names of table in database.

\item \code{list_TN_mysql}: to MySQL database and extract column names of table in database.

\item \code{list_VN_oracle}: to Oracle database and extract the column names of table in database.

\item \code{list_VN_mysql}: to MySQL database and extract column names of table in database.

\item \code{access_oracle}: to Oracle database

\item \code{access_mysql}: to MySQL database

\item \code{access_oracle_col}: to Oracle database and extract the column names of table in database.

\item \code{access_mysql_col}: to MySQL database and extract column names of table in database.
}}

