% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessData.R, R/accessDataCol.R
\name{access_ore}
\alias{access_ore}
\alias{access_ore_col}
\title{Functions to link to ORE database}
\usage{
access_ore(conn_string, database, table_name, username, password)

access_ore_col(conn_string, database, table_name, username, password)
}
\description{
Functions to link to ORE database

Functions to link to ORE database and extract column name
}
\section{Functions}{
\itemize{
\item \code{access_ore}: to ORE database

\item \code{access_ore_col}: to ORE database and extract column names of table in database
}}

