% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R
\name{list_TN_flat}
\alias{list_TN_flat}
\title{Functions to link to flat database and extract column name}
\usage{
list_TN_flat(wkdir, research.folder, database)
}
\description{
Functions to link to flat database and extract column name
}
