% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R
\name{list_TN_ore}
\alias{list_TN_ore}
\alias{list_TN_oracle}
\alias{list_TN_mysql}
\alias{list_TN_flat}
\title{Functions to link to ORE database and extract column name}
\usage{
list_TN_ore(database, conn_string, username, password)

list_TN_oracle(database, username, password)

list_TN_mysql(database, username, password)

list_TN_flat(wkdir, research.folder, database)
}
\arguments{
\item{database}{Name of database. Should be either \code{private} or
\code{public} for flat tables, indicating whether data is stored in
\code{public_data} or \code{research/[research folder]/private_data}, or
the actual name of database if extracting data from a database.}

\item{conn_string}{Connection string for accessing ORE server. Default is
\code{NA}.}

\item{username}{User name for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{password}{Password for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{wkdir}{The path to working directory. See \code{\link{initWkdir}} for
details on a working directory.}

\item{research.folder}{The name of research folder. See
\code{\link{initResearchFolder}} for details on a research folder.}
}
\description{
Link to ORE database and extract the column names of table in
  database.
}
\section{Functions}{
\itemize{
\item \code{list_TN_oracle}: Link to Oracle database and extract the column names
of table in database.

\item \code{list_TN_mysql}: Link to MySQL database and extract column names of
table in database.

\item \code{list_TN_flat}: Link to flat database and extract column names of
table.
}}

