% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R
\name{list_VN_mysql}
\alias{list_VN_mysql}
\alias{list_VN_flat}
\title{Functions to link to MySQL database and extract column name}
\usage{
list_VN_mysql(database, table_name, username, password)

list_VN_flat(wkdir, database, table_name)
}
\arguments{
\item{database}{Name of database. Should be either \code{private} or
\code{public} for flat tables, indicating whether data is stored in
\code{public_data} or \code{research/[research folder]/private_data}, or
the actual name of database if extracting data from a database.}

\item{table_name}{The name of dataset. File extension must be included if the
data is saved as a flat file instead of in a database, e.g. `data.csv`
should be given instead of simply `data`.}

\item{username}{User name for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{password}{Password for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{wkdir}{The path to working directory. See \code{\link{initWkdir}} for
details on a working directory.}
}
\description{
Link to MySQL database and extract column names of
  table in database.
}
\section{Functions}{
\itemize{
\item \code{list_VN_flat}: Link to flat tables and extract column names.
}}

