% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R
\name{list_VN_ore}
\alias{list_VN_ore}
\alias{list_VN_oracle}
\alias{list_VN_mysql}
\title{Functions to link to ORE database and extract column name}
\usage{
list_VN_ore(database, table_name, conn_string, username, password)

list_VN_oracle(database, table_name, username, password)

list_VN_mysql(database, table_name, username, password)
}
\arguments{
\item{database}{Name of database. Should be either \code{private} or
\code{public} for flat tables, indicating whether data is stored in
\code{public_data} or \code{research/[research folder]/private_data}, or
the actual name of database if extracting data from a database.}

\item{table_name}{The name of dataset. File extension must be included if the
data is saved as a flat file instead of in a database, e.g. `data.csv`
should be given instead of simply `data`.}

\item{conn_string}{Connection string for accessing ORE server. Default is
\code{NA}.}

\item{username}{User name for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{password}{Password for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}
}
\description{
Link to ORE database and extract the column names of table in
  database.
}
\details{
For flat files, must use \code{data.table::fread} to read
  deliminated files (i.e. \code{txt} and \code{csv}), because this will
  result in character ICD codes, which is consistent with ICD read by
  \code{readxl::read_excel} (used to read in selection).

  \code{xlsx::read.xlsx} will also read ICD code as character. So if
  \code{haven::read_dta} and \code{haven::read_sav} reads ICD as numeric,
  then we cannot do anything about it. (In simulated data it is still
  character)
}
\section{Functions}{
\itemize{
\item \code{list_VN_oracle}: Link to Oracle database and extract the column names
of table in database.

\item \code{list_VN_mysql}: Link to MySQL database and extract column names of
table in database.
}}

