% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R
\name{ListFN}
\alias{ListFN}
\title{List file name, table name and variable name of specific table
List all the file names of inclusion criteria or varaibles list  --------}
\usage{
ListFN(wkdir = getwd(), research.folder = NA, mode = NA)
}
\arguments{
\item{wkdir}{character string with the whole work-directory.}

\item{research.folder}{character string with the name of research folder's name}

\item{mode}{character string to indicate whether list file names of inclusion criteria file or variable list.}
}
\value{
character vector includes all the file names based on the mode choice.
}
\description{
List file name, table name and variable name of specific table
List all the file names of inclusion criteria or varaibles list  --------
}
