% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R
\name{ListTN}
\alias{ListTN}
\title{List all the table names the user can access --------}
\usage{
ListTN(wkdir = getwd(), database = NA, research.folder = NA,
  data.type = NA, conn_string = NA, username = NA, password = NA)
}
\arguments{
\item{wkdir}{character string with the whole work-directory}

\item{database}{character string with the real schema name or 'public'/'private'.}

\item{research.folder}{character string with the name of research folder, only use when your table saved in the private_data folder}

\item{data.type}{character string with data type including 'ore'/'oracle'/'sql'/'flat'.}

\item{conn_string}{character string only use when the data.type equal to 'ore'.}

\item{username}{character string only use when you need to access to server(i.e., ore,oracle,sql).}

\item{password}{character string only use when you need to access to server(i.e., ore,oracle,sql).}
}
\value{
character vector includes all the table names based on the database choice.
}
\description{
List all the table names the user can access --------
}
