% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R
\name{list_VN_flat}
\alias{list_VN_flat}
\title{Functions to link to flat database and extract column name}
\usage{
list_VN_flat(wkdir, database, table_name)
}
\arguments{
\item{wkdir}{The path to working directory. See \code{\link{initWkdir}} for
details on a working directory.}

\item{database}{Name of database. Should be either \code{private} or
\code{public} for flat tables, indicating whether data is stored in
\code{public_data} or \code{research/[research folder]/private_data}, or
the actual name of database if extracting data from a database.}

\item{table_name}{The name of dataset. File extension must be included if the
data is saved as a flat file instead of in a database, e.g. `data.csv`
should be given instead of simply `data`.}
}
\description{
Functions to link to flat database and extract column name
}
\section{Functions}{
\itemize{
\item \code{list_VN_flat}: oad R data (\code{.RData})
}}

