% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R, R/accessData.R, R/accessDataCol.R
\name{list_VN_flat}
\alias{list_VN_flat}
\alias{access_flat}
\alias{access_flat_col}
\title{Functions to link to flat database and extract column name}
\usage{
list_VN_flat(wkdir, database, table_name)

access_flat(database, table_name, data_type)

access_flat_col(database, table_name, data_type)
}
\arguments{
\item{table_name}{The name of file to read in, including file extension.}

\item{table_name}{The name of file to read in, including file extension.}

\item{table_name}{The name of file to read in, including file extension.}
}
\description{
Functions to link to flat database and extract column name

Functions to link to flat database

Functions to link to flat database and extract column name
}
\details{
Must use \code{data.table::fread} to read deliminated files (i.e.
  \code{txt} and \code{csv}), because this will result in character ICD
  codes, which is consistent with ICD read by \code{readxl::read_excel} (used
  to read in selection).

  \code{xlsx::read.xlsx} will also read ICD code as character. So if
  \code{haven::read_dta} and \code{haven::read_sav} reads ICD as numeric,
  then we cannot do anything about it. (In simulated data it is still
  character)

Request forms are read in with xlsx::read.xlsx, which reads in ICD
  codes as charactor. Functions to read in data should read ICD codes as
  character as well in optimal cases.

  Cannot use data.table::fread because it does not work with R version 3.2.0.

  \code{xlsx::read.xlsx} will also read ICD code as character. So if
  \code{haven::read_dta} and \code{haven::read_sav} reads ICD as numeric,
  then we cannot do anything about it. (In simulated data it is still
  character)
}
\section{Functions}{
\itemize{
\item \code{list_VN_flat}: R data (\code{.RData})

\item \code{access_flat}: R data (\code{.RData})

\item \code{access_flat_col}: column names of table (\code{character string})
}}

