% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_inclu}
\alias{merge_inclu}
\title{Find the intersection or union of all inclusion, based on \code{data_logic}}
\usage{
merge_inclu(inclu_list, data_logic)
}
\description{
Find the intersection or union of all inclusion, based on \code{data_logic}
}
