% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_var}
\alias{merge_var}
\title{Find the intersection or union of all variable lists, based on
\code{data_logic}}
\usage{
merge_var(var_list, data_logic)
}
\description{
Find the intersection or union of all variable lists, based on
\code{data_logic}
}
