% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_inclu}
\alias{process_inclu}
\title{Processing request form}
\usage{
process_inclu(research.folder, inclusion.xls.file, data_logic, overwrite,
  username = NA, password = NA)
}
\description{
Processing request form
}
