% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_var}
\alias{process_var}
\title{Process variable list}
\usage{
process_var(research.folder, variable.xls.file, inclu_list, overwrite,
  username = NA, password = NA)
}
\arguments{
\item{research.folder}{The name of research folder. See
\code{\link{initResearchFolder}} for details on a research folder.}

\item{variable.xls.file}{Name(s) of request form(s) with variable lists.
Multiple request forms should be specified as a character vector. Default
is \code{NA}, where no variable list is specified, and variables in the
inclusion criteria will be extracted instead.}

\item{inclu_list}{A list of inclusion to combine.}

\item{overwrite}{Whether to overwrite existing request form. Default is
\code{TRUE}.}

\item{username}{User name for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{password}{Password for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}
}
\description{
Process variable list
}
