% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDataXMan.R
\docType{package}
\name{RDataXMan-package}
\alias{RDataXMan-package}
\title{RDataXMan: Data Extraction Management}
\description{
There are 5 functions in this package.
}
\details{
Functions \code{\link{genInclusion}}, \code{\link{genVariable}} and
\code{\link{extract_data}} are the 3 main functions that generates inclusion
criteria and variable lists based on target databses, and extract subset of
data for user.

Functions \code{\link{initWkdir}} and \code{\link{initResearchFolder}} are
helper functions to set working directory and subfolders for specific data
requests.
}
