% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rename_cols}
\alias{rename_cols}
\title{Rename non-key variables in a list of variable names if they appear multiple
times in the list of variable names}
\usage{
rename_cols(name_list, all_list = name_list, key_list)
}
\description{
Rename non-key variables in a list of variable names if they appear multiple
times in the list of variable names
}
