% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{replace_key_var}
\alias{replace_key_var}
\title{Replaces `*` in a string with digits 0 to 9}
\usage{
replace_key_var(s)
}
\arguments{
\item{s}{A string.}
}
\description{
Replaces `*` in a string with digits 0 to 9
}
