% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R, R/accessData.R, R/accessDataCol.R
\name{success_msg}
\alias{success_msg}
\alias{success_msg}
\alias{success_msg}
\title{Flag message for successfully loading a table}
\usage{
success_msg(table_name, database, is_database = TRUE)

success_msg(table_name, database, is_database = TRUE)

success_msg(table_name, database, is_database = TRUE)
}
\description{
Flag message for successfully loading a table

Flag message for successfully loading a table

Flag message for successfully loading a table
}
