% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarise_cts}
\alias{summarise_cts}
\alias{summarise_cat}
\title{Summarise a continuous variable}
\usage{
summarise_cts(x)

summarise_cat(x)
}
\arguments{
\item{x}{Variable to summarise.}
}
\description{
Summarise a continuous variable
}
\section{Functions}{
\itemize{
\item \code{summarise_cat}: Summarise a continuous variable
}}

