% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessData.R
\name{time_to_char}
\alias{time_to_char}
\title{\code{ore.pull} automatically add time zone of Date type columns.}
\usage{
time_to_char(data)
}
\description{
\code{ore.pull} automatically add time zone of Date type columns.
}
