% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unzip}
\alias{unzip}
\title{Converts a comma-seperated string into a vector}
\usage{
unzip(string)
}
\arguments{
\item{string}{A comma-separated string.}
}
\description{
Converts a comma-seperated string into a vector
}
