% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_data_raw}
\alias{write_data_raw}
\title{Write a list of raw data to a group of \code{csv}}
\usage{
write_data_raw(data_list, type, research.folder)
}
\description{
Write a list of raw data to a group of \code{csv}
}
