% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_id}
\alias{write_id}
\title{Write the list of unique identifiers to \code{csv}}
\usage{
write_id(id_merged, type, research.folder)
}
\description{
Write the list of unique identifiers to \code{csv}
}
