% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_to_file}
\alias{write_to_file}
\title{Check whether file exists and write file}
\usage{
write_to_file(table, file, sheetName, overwrite = NULL)
}
\description{
Check whether file exists and write file
}
\details{
No need to check \code{overwrite} when adding sheets to the file. In
  this case, set it to \code{NULL}.
}
