% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessDataCol.R
\name{access_ore_col}
\alias{access_ore_col}
\alias{access_oracle_col}
\alias{access_mysql_col}
\alias{access_flat_col}
\title{Functions to link to ORE database and extract column name}
\usage{
access_ore_col(conn_string, database, table_name, username, password)

access_oracle_col(database, table_name, username, password)

access_mysql_col(database, table_name, username, password)

access_flat_col(database, table_name, data.type)
}
\arguments{
\item{conn_string}{Connection string for accessing ORE server. Default is
\code{NA}.}

\item{database}{Name of database. Should be either \code{private} or
\code{public} for flat tables, indicating whether data is stored in
\code{public_data} or \code{research/[research folder]/private_data}, or
the actual name of database if extracting data from a database.}

\item{table_name}{The name of dataset. File extension must be included if the
data is saved as a flat file instead of in a database, e.g. `data.csv`
should be given instead of simply `data`.}

\item{username}{User name for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{password}{Password for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{data.type}{\code{flat} for flat tables, \code{sql} fo rSQL database,
\code{ore} for Oracle R Enterprise (ORE) database, and \code{oracle} for
Oracle in SQL mode.}
}
\description{
Linking to ORE database and extract column names of table in
  database.
}
\section{Functions}{
\itemize{
\item \code{access_oracle_col}: Link to Oracle database and extract the column
names of table in database.

\item \code{access_mysql_col}: Link to MySQL database and extract column names of
table in database.

\item \code{access_flat_col}: Load column names of table (\code{character
string}).
}}

