% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_long}
\alias{check_long}
\title{Validate request form in long format}
\usage{
check_long(research.folder, database, request_file, template_file)
}
\arguments{
\item{research.folder}{The name of research folder. See
\code{\link{initResearchFolder}} for details on a research folder.}

\item{database}{Name of database. Should be either \code{private} or
\code{public} for flat tables, indicating whether data is stored in
\code{public_data} or \code{research/[research folder]/private_data}, or
the actual name of database if extracting data from a database.}

\item{request_file}{Inclusion list.}

\item{template_file}{Inclusion template generated by \code{genInclusion}.}
}
\description{
Validate request form in long format
}
\details{
\code{request_file} is annotated inclusion list, and
  \code{template_file} is the one generated by \code{genInclusion}. Column
  names and \code{key.var} should be identical in the two files.
}
