% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{factor_to_char}
\alias{factor_to_char}
\title{Converts factor columns to character}
\usage{
factor_to_char(data)
}
\arguments{
\item{data}{A data frame.}
}
\description{
Converts factor columns to character
}
\details{
This is needed because \code{ore.pull} automatically converts
  character columns to factors.
}
