% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R
\name{ListFN}
\alias{ListFN}
\title{List file name, table name and variable name of specific table}
\usage{
ListFN(wkdir = getwd(), research.folder = NA, mode = NA)
}
\arguments{
\item{wkdir}{The path to working directory. See \code{\link{initWkdir}} for
details on a working directory.}

\item{research.folder}{The name of research folder. See
\code{\link{initResearchFolder}} for details on a research folder.}

\item{mode}{character string to indicate whether list file names of inclusion
criteria file or variable list.}
}
\value{
character vector includes all the file names based on the mode
  choice.
}
\description{
List all the file names of inclusion criteria or variables list.
}
