% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListName.R
\name{ListVN}
\alias{ListVN}
\title{List all the variable names for specific table}
\usage{
ListVN(
  wkdir = getwd(),
  research.folder = NA,
  data.type = NA,
  database = NA,
  table_name = NA,
  conn_string = NA,
  username = NA,
  password = NA
)
}
\arguments{
\item{wkdir}{The path to working directory. See \code{\link{initWkdir}} for
details on a working directory.}

\item{research.folder}{The name of research folder. See
\code{\link{initResearchFolder}} for details on a research folder.}

\item{data.type}{\code{flat} for flat tables and \code{sql} fo MySQL database.}

\item{database}{Name of database. Should be either \code{private} or
\code{public} for flat tables, indicating whether data is stored in
\code{public_data} or \code{research/[research folder]/private_data}, or
the actual name of database if extracting data from a database.}

\item{table_name}{For database this is the name of table. For flat file this
is the name of file to read in, including file extension.}

\item{username}{User name for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{password}{Password for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}
}
\value{
character vector includes all the table names based on the database
  choice.
}
\description{
List all the variable names for specific table
}
