% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessData.R
\name{access_flat}
\alias{access_flat}
\title{Functions to link to flat database}
\usage{
access_flat(database, table_name, data.type)
}
\arguments{
\item{database}{Name of database. Should be either \code{private} or
\code{public} for flat tables, indicating whether data is stored in
\code{public_data} or \code{research/[research folder]/private_data}, or
the actual name of database if extracting data from a database.}

\item{table_name}{The name of dataset. File extension must be included if the
data is saved as a flat file instead of in a database, e.g. `data.csv`
should be given instead of simply `data`.}

\item{data.type}{\code{flat} for flat tables and \code{sql} fo MySQL database.}
}
\description{
Functions to link to flat database
}
\details{
Request forms are read in with xlsx::read.xlsx, which reads in ICD
  codes as charactor. Functions to read in data should read ICD codes as
  character as well in optimal cases.

  Cannot use data.table::fread because it does not work with R version 3.2.0.

  \code{xlsx::read.xlsx} will also read ICD code as character. So if
  \code{haven::read_dta} and \code{haven::read_sav} reads ICD as numeric,
  then we cannot do anything about it. (In simulated data it is still
  character)
}
\section{Functions}{
\itemize{
\item \code{access_flat}: R data (\code{.RData})
}}

