% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_inclu}
\alias{merge_inclu}
\title{Find the intersection or union of all inclusion, based on \code{dataLogic}}
\usage{
merge_inclu(inclu_list, dataLogic)
}
\arguments{
\item{inclu_list}{A list of inclusion to combine.}

\item{dataLogic}{Whether to take \code{union} or \code{intersection} of
inclusion criteria, if multiple criteria specified. Default is \code{NA}.}
}
\description{
Find the intersection or union of all inclusion, based on \code{dataLogic}
}
