% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_inclu}
\alias{process_inclu}
\title{Processing request form}
\usage{
process_inclu(
  research.folder,
  inclusion.xls.file,
  dataLogic,
  overwrite,
  username = NA,
  password = NA
)
}
\arguments{
\item{research.folder}{The name of research folder. See
\code{\link{initResearchFolder}} for details on a research folder.}

\item{inclusion.xls.file}{Name(s) of request form(s) with inclusion
criterion. Multiple request forms should be specified as a character
vector.}

\item{dataLogic}{Whether to take \code{union} or \code{intersection} of
inclusion criteria, if multiple criteria specified. Default is \code{NA}.}

\item{overwrite}{Whether to overwrite existing request form. Default is
\code{TRUE}.}

\item{username}{User name for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}

\item{password}{Password for accessing database if \code{data.type} is not
\code{flat}. Default is \code{NA} for flat tables.}
}
\description{
Processing request form
}
