% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{success_msg}
\alias{success_msg}
\title{Flag message for successfully loading a table}
\usage{
success_msg(table_name, database, is_database = TRUE)
}
\arguments{
\item{table_name}{The name of dataset. File extension must be included if the
data is saved as a flat file instead of in a database, e.g. `data.csv`
should be given instead of simply `data`.}

\item{database}{Name of database. Should be either \code{private} or
\code{public} for flat tables, indicating whether data is stored in
\code{public_data} or \code{research/[research folder]/private_data}, or
the actual name of database if extracting data from a database.}

\item{is_database}{Whether table is in a database.}
}
\description{
Flag message for successfully loading a table
}
