% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{time_to_char}
\alias{time_to_char}
\title{Convert time to character}
\usage{
time_to_char(data)
}
\arguments{
\item{data}{A data frame.}
}
\description{
Convert time to character
}
\details{
This is because \code{ore.pull} automatically add time zone of Date
  type columns.
}
